feat(types): export all consumer-facing types from core and server packages
- Export 8 internal types from @jaicome/zatca-core simplified_tax_invoice_template
- Add 4 workflow result types to @jaicome/zatca-server (OnboardResult, SingleReportResult, BatchReportResult, ClientInvoiceRecord)
- Update example files to import types from packages instead of defining locally
- Add skipLibCheck to zatca-server tsconfig to workaround pre-existing dependency type errors
