+++ packages/zatca-server/src/api/index.ts
@@ -319,4 +319,8 @@ export type {
   InvoiceResponse,
   IssuedCertificate,
   ValidationMessage,
+  OnboardResult,
+  SingleReportResult,
+  BatchReportResult,
+  ClientInvoiceRecord,
 } from "./types.js";
